requirements.txt
pandas
numpy
scikit-learn
nltk
pyyaml
matplotlib
seaborn
joblib
